+++ src/platform/packages/shared/kbn-connector-specs/src/specs/alienvault_otx/icon/alienvault.jpg
