+++ src/platform/packages/shared/kbn-connector-specs/src/specs/alienvault_otx/icon/index.tsx
@@ -7,12 +7,13 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
+/* eslint-disable import/no-default-export */
+
 import React from 'react';
 
 import { EuiIcon } from '@elastic/eui';
 import type { ConnectorIconProps } from '../../../types';
-
-import icon from '../../../icons/alienvault_otx/alienvault.jpg';
+import icon from './alienvault.jpg';
 
 export default (props: ConnectorIconProps) => {
   return <EuiIcon type={icon} {...props} />;
