+++ src/platform/packages/shared/kbn-connector-specs/src/specs/virustotal/icon/virustotal.png
